Add files via upload
This was a test back in my "Microinformatics Programming" class at the university right in the first semester. We used "Access" to make a simple database of an fictional confectionary called "SweetTec". We applied techniques such as CRUD to make this application funcional. It was a lot of fun creating this project since it was the first time we actually could see a face of a system of us working.